pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gui-agent-harness"
version = "0.3.0"
description = "GUI automation powered by Agentic Programming — designed for OpenClaw"
requires-python = ">=3.11"
dependencies = [
    "agentic-programming @ file:libs/agentic-programming",
    "pynput",
    "opencv-python",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30"]
openai = ["openai>=1.30"]
dev = [
    "pytest>=8.0",
    "numpy",
]

[tool.setuptools.packages.find]
include = ["gui_harness*"]

# Note: install agentic-programming submodule first:
#   pip install -e ./libs/agentic-programming
